// src/models/ImportItemSave.ts
export interface ImportItemSave {
    id?: number; //auto increment
    tin: string;
    bhfId: string;
    taskCd: string;
    dclDe: string;
    itemSeq: number;
    hsCd: string;
    itemClCd: string;
    itemCd: string;
    imptItemsttsCd: string;
    remark?: string;
    modrNm: string;
    modrId: string;
}